### 2018-07-24 #834 @whedon
[PRE REVIEW]: fgivenx: a python package for functional posterior plotting
**Submitting author:** @williamjameshandley (<a href="http://orcid.org/0000-0002-5866-0445">Will Handley</a>)
**Repository:** <a href="https://github.com/williamjameshandley/fgivenx" target ="_blank">https://github.com/williamjameshandley/fgivenx</a>
**Version:** v2.1.0
