### 2018-07-24 #834 @arfon
:wave: @williamjameshandley - could you please take a look at this list of potential reviewers and identify some potential candidates to review your submission? https://bit.ly/joss-reviewers
