### 2018-07-24 #834 @arfon
:wave: @zhampel @FaustinCarter - would you be willing to review this submission for JOSS?
