### 2018-07-25 #834 @FaustinCarter
Sure. I can do that. Give me a week or so to look over the code. After skimming the paper quickly I think it could benefit from a rewrite. Some specific comments by paragraph:

1:
-Why does this end in a colon? Are you about to present an example?
-Your tense is off. I think you mean "probability distributions" rather than "a probability distribution" (which would imply a single distribution for all unknown parameters in the universe!)

2:
-I'm not sure what the point of this paragraph is. The age of the universe and the idea that space expands don't seem to be relevant at any further point in the paper.

3:
-Plots like what? Are you suggesting that the idea in the above paragraph could be parameterized into two variables that have a joint probability distribution? What are those variables? Can you include the plot of those variables made from your software?

4:
-scipy, getdist, and corner all have instructions on how to cite them in an academic publication. It would be good to include those citations in addition to links to their webpages. A reference to a MCMC paper would also be good to make it easy for readers to look that up.
-I'll also put in a shameless plug here for the pygtc package (10.21105/joss.00046) which is an alternative to corner and probably belongs in that list, although that's up to you.
-After a more careful reading I think you are implying that corner plots are nice, but there is more to the story (i.e. looking at the actual model described by all the MCMC realizations?) and that's where your software comes in? Might be good to make this more explicit.

5:
-Mixed voice. You switch between third person (one does, one could) and first person plural (we do, we could, our ability). It would be good to unify the voice over the whole paper.
-Is the phrase "plotted thus" supposed to be a reference to Fig. 1? If so, say something like: "in Fig. 1 panel (a) we plot the parameter covariances between m and c for X realizations. In panel (b) we plot the function y=mx+c for X realizations. In panel (c) we plot something called DxL which is never mentioned in the paper and should be. In panel (d) we plot the probability of measuring y for a given x, which is essential a contour version of panel (b)."
-Why is there a red distribution and a blue distribution? Is that two separate sets of fake data that are being fit to the same model?

6:
-A couple references seem broken?

7:
-The first line of this is the first mention you make of what your package actually does in words. I think this should come earlier and there should be a little more description of what fgivenx is designed to do right from the start.

That's enough for now. Let me dig into the code an run it and I'll have more comments. Oh, and is there a link to hosted documentation somewhere (I didn't see one on the github page)? Preferably documentation that includes installation instructions and a description of options (see the docs for corner as a great example)?
