### 2018-07-27 #849 @FaustinCarter
I have some specific comments on the paper which I outline here by paragraph. I'll have comments on the code soon.

Paragraph 1:
------------
 - Why does this end in a colon? Are you about to present an example?
 - Your tense is off. I think you mean "probability distributions" rather than "a probability distribution" (which would imply a single distribution for all unknown parameters in the universe!)

Paragraph 2:
-------------
 - I'm not sure what the point of this paragraph is. The age of the universe and the idea that space expands don't seem to be relevant at any further point in the paper.

Paragraph 3:
-------------
 - Plots like what? Are you suggesting that the idea in the above paragraph could be parameterized into two variables that have a joint probability distribution? What are those variables? Can you include the plot of those variables made from your software?

Paragraph 4:
-------------
 - `scipy`, `getdist`, and `corner` all have instructions on how to cite them in an academic publication. It would be good to include those citations in addition to links to their webpages. A reference to a MCMC paper would also be good to make it easy for readers to look that up.
 - I'll also put in a shameless plug here for the `pygtc` package ([10.21105/joss.00046](http://joss.theoj.org/papers/fac5017b39292e539933d12936615f4a)) which is an alternative to `corner` and probably belongs in that list, although that's up to you.
 - After a more careful reading I think you are implying that corner plots are nice, but there is more to the story (i.e. looking at the actual model described by all the MCMC realizations?) and that's where your software comes in? I think you need to make this more explicit.

Paragraph 5:
-------------
 - Mixed voice. You switch between third person (one does, one could) and first person plural (we do, we could, our ability). It would be good to unify the voice over the whole paper.
 - Is the phrase "plotted thus" supposed to be a reference to Fig. 1? If so, say something like: "in Fig. 1 panel (a) we plot the parameter covariances between _m_ and _c_ for _X_ realizations (blue) and _Y_ realizations (red). In panel (b) we plot the function _y_=_mx_+_c_ for _X_ realizations. In panel (c) we plot something called _DxL_ which is never mentioned in the paper and should be. In panel (d) we plot the probability of measuring _y_ for a given _x_, which is essential a contour version of panel (b)."
 - Why is there a red distribution and a blue distribution? Is that two separate sets of fake data that are being fit to the same model?

Paragraph 6:
-------------
 - A couple references seem broken.

Paragraph 7:
-------------
 - The first line of this is the first mention you make of what your package actually does in words. I think this should come earlier and there should be a little more description of what `fgivenx` is designed to do right from the start.
