### 2018-07-29 #849 @FaustinCarter
A few comments on the install process:

 - I could not find installation instructions anywhere ( so I just guessed with pip). I also could not find a link to the documentation? Is it hosted somewhere, or is the user supposed to build it themselves? If the second, it would be good to include installation instructions and document building instructions directly in the github readme and in the description on PyPI.

 - No description of anything at all on the PyPI page for fgivenx.

 - For some reason the tests directory is not included in the source that is available at PyPI (this is v. 2.1.0), which means the tests do not install with pip. They are available in the zipped up source files in the github release.
