### 2018-08-03 20f084d
Added JOSS link to README.md
