### 2018-08-03 839a003
Added tests to setup.py
