### 2018-08-03 06fe526
Updated tests to check for getdist imports
