+++ fgivenx/tests/test_samples.py
@@ -4,7 +4,60 @@ import os
 from shutil import rmtree
 from numpy.testing import assert_allclose, assert_array_equal
 from fgivenx.samples import compute_samples, samples_from_getdist_chains
-import getdist
+
+try:
+    import getdist
+
+    def test_samples_from_getdist_chains():
+
+        # Set up getdist chains
+        file_root = './chains/test'
+        labels = [r'\alpha', r'\beta', r'\gamma']
+        names = ['a', 'b', 'g']
+        chains_file = file_root + '.txt'
+        pars_file = file_root + '.paramnames'
+        nsamples = 1000
+        params = ['a', 'g']
+        i = [names.index(p) for p in params]
+        numpy.random.seed(0)
+        samples_ = numpy.random.rand(nsamples, len(names))
+        weights_ = numpy.random.rand(nsamples)
+        samples = getdist.mcsamples.MCSamples(samples=samples_, labels=labels,
+                                              names=names, weights=weights_)
+        samples.saveAsText(file_root, make_dirs=True)
+
+        samples, weights = samples_from_getdist_chains(params, file_root=file_root)
+        assert_allclose(samples, samples_[:, i])
+        assert_allclose(weights, weights_)
+
+        # now test function
+        with pytest.raises(ValueError):
+            samples_from_getdist_chains(params)
+        with pytest.raises(ValueError):
+            samples_from_getdist_chains(params, chains_file=chains_file)
+        with pytest.raises(ValueError):
+            samples_from_getdist_chains(params, paramnames_file=pars_file)
+
+        samples, weights = samples_from_getdist_chains(params,
+                                                       chains_file=chains_file,
+                                                       paramnames_file=pars_file)
+        assert_allclose(samples, samples_[:, i])
+        assert_allclose(weights, weights_)
+
+        samples, weights, latex = samples_from_getdist_chains(params,
+                                                              file_root=file_root,
+                                                              latex=True)
+        assert_allclose(weights, weights_)
+        assert_array_equal(latex, numpy.array(labels)[i])
+
+        with open(chains_file, "w"):
+            pass
+        samples, weights = samples_from_getdist_chains(params, file_root=file_root)
+
+        rmtree('chains')
+
+except ImportError:
+    pass
 
 
 def test_compute_samples():
@@ -39,50 +92,3 @@ def test_compute_samples():
     rmtree('.test_cache')
 
 
-def test_samples_from_getdist_chains():
-
-    # Set up getdist chains
-    file_root = './chains/test'
-    labels = [r'\alpha', r'\beta', r'\gamma']
-    names = ['a', 'b', 'g']
-    chains_file = file_root + '.txt'
-    pars_file = file_root + '.paramnames'
-    nsamples = 1000
-    params = ['a', 'g']
-    i = [names.index(p) for p in params]
-    numpy.random.seed(0)
-    samples_ = numpy.random.rand(nsamples, len(names))
-    weights_ = numpy.random.rand(nsamples)
-    samples = getdist.mcsamples.MCSamples(samples=samples_, labels=labels,
-                                          names=names, weights=weights_)
-    samples.saveAsText(file_root, make_dirs=True)
-
-    samples, weights = samples_from_getdist_chains(params, file_root=file_root)
-    assert_allclose(samples, samples_[:, i])
-    assert_allclose(weights, weights_)
-
-    # now test function
-    with pytest.raises(ValueError):
-        samples_from_getdist_chains(params)
-    with pytest.raises(ValueError):
-        samples_from_getdist_chains(params, chains_file=chains_file)
-    with pytest.raises(ValueError):
-        samples_from_getdist_chains(params, paramnames_file=pars_file)
-
-    samples, weights = samples_from_getdist_chains(params,
[…]
