### 2018-08-03 6bb8f43
Updates to the paper after the first round of review
