+++ paper/paper.bib
@@ -1,104 +1,180 @@
-@ARTICLE{Inflation2015,
-   author = {{Planck Collaboration}},
+@ARTICLE{inflation2015,
+    author = {{Planck Collaboration}},
     title = "{Planck 2015 results. XX. Constraints on inflation}",
-  journal = {\aap},
-archivePrefix = "arXiv",
-   eprint = {1502.02114},
- keywords = {cosmic background radiation, cosmology: theory, early Universe, inflation},
-     year = 2016,
+    journal = {\aap},
+    archivePrefix = "arXiv",
+    eprint = {1502.02114},
+    keywords = {cosmic background radiation, cosmology: theory, early Universe, inflation},
+    year = 2016,
     month = sep,
-   volume = 594,
-      eid = {A20},
+    volume = 594,
+    eid = {A20},
     pages = {A20},
-      doi = {10.1051/0004-6361/201525898},
-   adsurl = {http://adsabs.harvard.edu/abs/2016A%26A...594A..20P},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    doi = {10.1051/0004-6361/201525898},
+    adsurl = {http://adsabs.harvard.edu/abs/2016A%26A...594A..20P},
+    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+
+@ARTICLE{legacy2018,
+    author = {{Planck Collaboration}},
+    title = "{Planck 2018 results. I. Overview and the cosmological legacy of Planck}",
+    journal = {ArXiv e-prints},
+    archivePrefix = "arXiv",
+    eprint = {1807.06205},
+    keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
+    year = 2018,
+    month = jul,
+    adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180706205P},
+    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@ARTICLE{inflation2018,
+    author = {{Planck Collaboration}},
+    title = "{Planck 2018 results. X. Constraints on inflation}",
+    journal = {ArXiv e-prints},
+    archivePrefix = "arXiv",
+    eprint = {1807.06211},
+    keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
+    year = 2018,
+    month = jul,
+    adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180706211P},
+    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
 @ARTICLE{Hee2015,
-   author = {{Hee}, S. and {Handley}, W.~J. and {Hobson}, M.~P. and {Lasenby}, A.~N.
-	},
+    author = {{Hee}, S. and {Handley}, W.~J. and {Hobson}, M.~P. and {Lasenby}, A.~N.
+    },
     title = "{Bayesian model selection without evidences: application to the dark energy equation-of-state}",
-  journal = {\mnras},
-archivePrefix = "arXiv",
-   eprint = {1506.09024},
- keywords = {equation of state, methods: data analysis, methods: statistical, cosmological parameters, dark energy},
-     year = 2016,
+    journal = {\mnras},
+    archivePrefix = "arXiv",
+    eprint = {1506.09024},
+    keywords = {equation of state, methods: data analysis, methods: statistical, cosmological parameters, dark energy},
+    year = 2016,
     month = jan,
-   volume = 455,
+    volume = 455,
     pages = {2461-2473},
-      doi = {10.1093/mnras/stv2217},
-   adsurl = {http://adsabs.harvard.edu/abs/2016MNRAS.455.2461H},
[…]
