### 2018-08-03 87f59b3
Updates to README.md after review
