### 2018-08-03 #849 @williamjameshandley
@FaustinCarter

My apologies for the poor initial state of the paper, I'm new to this system, and it wasn't clear to me initially how to preview the paper rendering prior to submission.

Re paras 2 and 3, there was a missing figure, the presence of which hopefully makes them more clear.

I believe I have addressed all of the issues you raised, but it wasn't clear to me how to cite getdist
