### 2018-08-03 #849 @williamjameshandley
@FaustinCarter
>* I could not find installation instructions anywhere ( so I just guessed with pip). I also could not find a link to the documentation? Is it hosted somewhere, or is the user supposed to build it themselves? If the second, it would be good to include installation instructions and document building instructions directly in the github readme and in the description on PyPI.
>* No description of anything at all on the PyPI page for fgivenx.
>* For some reason the tests directory is not included in the source that is available at PyPI (this is v. 2.1.0), which means the tests do not install with pip. They are available in the zipped up source files in the github release.

I've updated the README.md, and now updated the pypi so that it mirrors it, you can see it in alpha state [here](https://pypi.org/project/fgivenx/2.1.1a/). Tests now included in pypi.

> The tests passed except for one which failed when it tried to import getdist. I think that since getdist is not a dependency of fgivenx, the tests should not fail if it is not installed. Instead, the code should check to see if getdist is available and skip the test in question if the import fails.

I've now done this, but I'm not sure in the best way to implement this (In particular, it broke my 100% code coverage).
