### 2018-08-03 e437854
Updated README.md citation
