+++ README.md
@@ -52,8 +52,8 @@ make html
 If you use fgivenx to generate plots for a publication, please cite as:
 ```bibtex
 @article{fgivenx,
-    doi = {xx.xxxxx/joss.xxxxx},
-    url = {http://dx.doi.org/xx.xxxxx/joss.xxxxx},
+    doi = {10.21105/joss.00849},
+    url = {http://dx.doi.org/10.21105/joss.00849},
     year  = {xxxx},
     month = {xxx},
     publisher = {The Open Journal},
@@ -63,7 +63,6 @@ If you use fgivenx to generate plots for a publication, please cite as:
     title = {fgivenx: Functional Posterior Plotter},
     journal = {The Journal of Open Source Software}
 }
-
 ```
 
 ## Example Usage
