### 2018-08-03 f60ea3b
Further formatting changes to README.md
