### 2018-08-04 0badc18
Converted readme to rst
