+++ README.md
@@ -1,163 +1,0 @@
-[![Build Status](https://travis-ci.org/williamjameshandley/fgivenx.svg?branch=master)](https://travis-ci.org/williamjameshandley/fgivenx)
-[![codecov](https://codecov.io/gh/williamjameshandley/fgivenx/branch/master/graph/badge.svg)](https://codecov.io/gh/williamjameshandley/fgivenx)
-[![PyPI version](https://badge.fury.io/py/fgivenx.svg)](https://badge.fury.io/py/fgivenx)
-[![Documentation Status](https://readthedocs.org/projects/fgivenx/badge/?version=latest)](https://fgivenx.readthedocs.io/en/latest/?badge=latest)
-[![status](http://joss.theoj.org/papers/cf6f8ac309d6a18b6d6cf08b64aa3f62/status.svg)](http://joss.theoj.org/papers/cf6f8ac309d6a18b6d6cf08b64aa3f62)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1310715.svg)](https://doi.org/10.5281/zenodo.1310715)
-
-# fgivenx: Functional Posterior Plotter
-![](https://raw.github.com/williamjameshandley/fgivenx/master/plot.png)
-
-These packages allow one to compute a predictive posterior of a function,
-dependent on sampled parameters. We assume one has a Bayesian posterior
-`Post(theta|D,M)` described by a set of posterior samples `{theta_i}~Post`. If
-there is a function parameterised by theta `f(x;theta)`, then this script will
-produce a contour plot of the conditional posterior `P(f|x,D,M)` in the `(x,f)`
-plane.
-
-The driving routine is `fgivenx.compute_contours`, and example usage can be
-found by running `help(fgivenx)`. It is compatible with getdist, and has a
-loading function provided by `fgivenx.samples.samples_from_getdist_chains`.
-
-
-
-## Installation
-
-Users can install using pip:
-```bash
-pip install fgivenx
-```
-using the setup.py:
-```bash
-git clone https://github.com/williamjameshandley/fgivenx
-cd fgivenx
-python setup.py install --user
-```
-or for those on [Arch linux](https://www.archlinux.org/) it is available on the
-[AUR](https://aur.archlinux.org/packages/python-fgivenx/)
-
-
-## Documentation
-
-Documentation is hosted at [ReadTheDocs](http://fgivenx.readthedocs.io/).
-
-To build your own local copy of the documentation you'll need to install
-[sphinx](http://www.sphinx-doc.org/en/master/). You can then run: 
-```bash
-cd docs
-make html
-``` 
-
-## Citation
-
-If you use `fgivenx` to generate plots for a publication, please cite as:
-```bibtex
-@article{fgivenx,
-    doi = {10.21105/joss.00849},
-    url = {http://dx.doi.org/10.21105/joss.00849},
-    year  = {xxxx},
-    month = {xxx},
-    publisher = {The Open Journal},
-    volume = {X},
-    number = {X},
-    author = {Will Handley},
-    title = {fgivenx: Functional Posterior Plotter},
-    journal = {The Journal of Open Source Software}
-}
-```
-
-## Example Usage
-
-```python
-import numpy
-import matplotlib.pyplot as plt
-from fgivenx import compute_samples, compute_pmf, compute_dkl
-from fgivenx.plot import plot, plot_lines
-
-
[…]
