+++ README.rst
@@ -1,0 +1,192 @@
+.. image:: https://travis-ci.org/williamjameshandley/fgivenx.svg?branch=master
+   :target: https://travis-ci.org/williamjameshandley/fgivenx
+   :alt: Build Status
+.. image:: https://codecov.io/gh/williamjameshandley/fgivenx/branch/master/graph/badge.svg
+   :target: https://codecov.io/gh/williamjameshandley/fgivenx
+   :alt: Test Coverage Status
+.. image:: https://badge.fury.io/py/fgivenx.svg
+   :target: https://badge.fury.io/py/fgivenx
+   :alt: PyPi location
+.. image:: https://readthedocs.org/projects/fgivenx/badge/?version=latest
+   :target: https://fgivenx.readthedocs.io/en/latest/?badge=latest
+   :alt: Documentation Status
+.. image:: http://joss.theoj.org/papers/cf6f8ac309d6a18b6d6cf08b64aa3f62/status.svg
+   :target: http://joss.theoj.org/papers/cf6f8ac309d6a18b6d6cf08b64aa3f62
+   :alt: Review Status
+.. image:: https://zenodo.org/badge/100947684.svg
+   :target: https://zenodo.org/badge/latestdoi/100947684
+   :alt: Permanent DOI
+
+
+fgivenx: Functional Posterior Plotter
+=====================================
+
+|image0|
+
+These packages allow one to compute a predictive posterior of a
+function, dependent on sampled parameters. We assume one has a Bayesian
+posterior ``Post(theta|D,M)`` described by a set of posterior samples
+``{theta_i}~Post``. If there is a function parameterised by theta
+``f(x;theta)``, then this script will produce a contour plot of the
+conditional posterior ``P(f|x,D,M)`` in the ``(x,f)`` plane.
+
+The driving routine is ``fgivenx.compute_contours``, and example usage
+can be found by running ``help(fgivenx)``. It is compatible with
+getdist, and has a loading function provided by
+``fgivenx.samples.samples_from_getdist_chains``.
+
+Installation
+------------
+
+Users can install using pip:
+
+.. code:: bash
+
+   pip install fgivenx
+
+using the setup.py:
+
+.. code:: bash
+
+   git clone https://github.com/williamjameshandley/fgivenx
+   cd fgivenx
+   python setup.py install --user
+
+or for those on `Arch linux <https://www.archlinux.org/>`__ it is
+available on the
+`AUR <https://aur.archlinux.org/packages/python-fgivenx/>`__
+
+Documentation
+-------------
+
+Documentation is hosted at
+`ReadTheDocs <http://fgivenx.readthedocs.io/>`__.
+
+To build your own local copy of the documentation you’ll need to install
+`sphinx <http://www.sphinx-doc.org/en/master/>`__. You can then run:
+
+.. code:: bash
+
+   cd docs
+   make html
+
+Citation
+--------
+
+If you use ``fgivenx`` to generate plots for a publication, please cite
+as:
+
+.. code:: bibtex
+
+   @article{fgivenx,
+       doi = {10.21105/joss.00849},
+       url = {http://dx.doi.org/10.21105/joss.00849},
+       year  = {xxxx},
+       month = {xxx},
+       publisher = {The Open Journal},
+       volume = {X},
+       number = {X},
+       author = {Will Handley},
+       title = {fgivenx: Functional Posterior Plotter},
[…]
