+++ setup.py
@@ -1,20 +1,19 @@
 #!/usr/bin/env python3
-
 from distutils.core import setup
-try:
-    import pypandoc
-    long_description = pypandoc.convert_file('README.md', 'rst')
-except (IOError, ImportError):
-    long_description = ''
+
+def readme():
+    with open('README.rst') as f:
+        return f.read()
+
 
 setup(name='fgivenx',
-      version='2.1.1a',
+      version='2.1.2a',
       author='Will Handley',
       author_email='wh260@cam.ac.uk',
       url='https://github.com/williamjameshandley/fgivenx',
       packages=['fgivenx', 'fgivenx.tests'],
       install_requires=['numpy','matplotlib','scipy','joblib','tqdm'],
-      tests_require=['pytest'],
+      tests_require=['pytest', 'pytest-cov'],
       license='MIT',
       classifiers=[
       'Development Status :: 4 - Beta',
@@ -28,6 +27,6 @@ setup(name='fgivenx',
       'Topic :: Scientific/Engineering :: Visualization',
       'Topic :: Scientific/Engineering :: Information Analysis',
       ],
-      description='Functional Posterior Plotter',
-      long_description=long_description,
+      description='fgivenx: Functional Posterior Plotter',
+      long_description=readme()
       )
