+++ requirements.txt
@@ -4,8 +4,6 @@ pytest
 scipy
 numpy
 joblib
-pypandoc
 numpydoc
-codecov
 getdist
 pytest-cov
