### 2018-08-04 f13e778
Linked documentation to README.rst
