### 2018-08-04 cf36bdd
Tests with minimal requirements
