+++ fgivenx/parallel.py
@@ -1,5 +1,29 @@
-import tqdm
-from joblib import Parallel, delayed, cpu_count
+try:
+    from tqdm import tqdm, tqdm_notebook
+except ImportError:
+    def tqdm(x, **kwargs):
+        return x
+
+    def tqdm_notebook(x, **kwargs):
+        return x
+
+
+try:
+    from joblib import Parallel, delayed, cpu_count
+except ImportError:
+    class Parallel(object):
+        def __init__(self, n_jobs=None):
+            pass
+
+        def __call__(self, x):
+            return list(x)
+
+    def delayed(x):
+        return x
+
+
+    def cpu_count():
+        return 1
 
 
 def parallel_apply(f, array, **kwargs):
@@ -46,10 +70,10 @@ def parallel_apply(f, array, **kwargs):
         raise TypeError('Unexpected **kwargs: %r' % kwargs)
     try:
         # If running in a jupyter notebook then use tqdm_notebook.
-        if get_ipython().has_trait('kernel'): progress = tqdm.tqdm_notebook
+        if get_ipython().has_trait('kernel'): progress = tqdm_notebook
     except (NameError, AssertionError):
         # Otherwise use regular tqdm progress bar
-        progress = tqdm.tqdm
+        progress = tqdm
     if not parallel:
         return [f(*(precurry + (x,) + postcurry)) for x in
                 progress(array, **tqdm_kwargs)]
