+++ .environment.yml
@@ -2,14 +2,16 @@ name: fgivenx-read-the-docs-environment
 
 dependencies:
     - python=3.6
+    - python=3.5
+    - python=3.4
     - python=2.7
     - matplotlib
-    - tqdm
-    - pytest
     - scipy
     - numpy
-    - joblib
     - pypandoc
     - numpydoc
-    - codecov
+    - getdist
+    - joblib
+    - tqdm
+    - pytest
     - pytest-cov
