### 2018-08-04 23c4c23
Added back coverage dependency
