+++ setup.py
@@ -5,9 +5,11 @@ try:
 except ImportError:
     from distutils.core import setup
 
+
 def readme():
     with open('README.rst') as f:
         return f.read()
+
 
 def get_version(short=False):
     with open('README.rst') as f:
@@ -20,6 +22,7 @@ def get_version(short=False):
                 else:
                     return ver
 
+
 setup(name='fgivenx',
       version=get_version(),
       description='fgivenx: Functional Posterior Plotter',
@@ -28,19 +31,25 @@ setup(name='fgivenx',
       author_email='wh260@cam.ac.uk',
       url='https://github.com/williamjameshandley/fgivenx',
       packages=['fgivenx', 'fgivenx.tests'],
-      install_requires=['numpy','matplotlib','scipy','joblib','tqdm'],
-      tests_require=['pytest', 'pytest-cov'],
+      install_requires=['matplotlib', 'numpy', 'scipy'],
+      extra_requires=['joblib', 'tqdm'],
+      tests_require=['pytest', 'pytest-cov', 'coverage'],
       license='MIT',
       classifiers=[
-      'Development Status :: 4 - Beta',
-      'Intended Audience :: Developers',
-      'Natural Language :: English',
-      'License :: OSI Approved :: MIT License',
-      'Programming Language :: Python :: 2.7',
-      'Programming Language :: Python :: 3.6',
-      'Topic :: Scientific/Engineering :: Astronomy',
-      'Topic :: Scientific/Engineering :: Physics',
-      'Topic :: Scientific/Engineering :: Visualization',
-      'Topic :: Scientific/Engineering :: Information Analysis',
+                   'Development Status :: 5 - Production/Stable',
+                   'Intended Audience :: Developers',
+                   'Intended Audience :: Science/Research',
+                   'Natural Language :: English',
+                   'License :: OSI Approved :: MIT License',
+                   'Programming Language :: Python :: 2.7',
+                   'Programming Language :: Python :: 3.4',
+                   'Programming Language :: Python :: 3.5',
+                   'Programming Language :: Python :: 3.6',
+                   'Topic :: Scientific/Engineering',
+                   'Topic :: Scientific/Engineering :: Astronomy',
+                   'Topic :: Scientific/Engineering :: Physics',
+                   'Topic :: Scientific/Engineering :: Visualization',
+                   'Topic :: Scientific/Engineering :: Information Analysis',
+                   'Topic :: Scientific/Engineering :: Mathematics',
       ],
       )
