+++ README.rst
@@ -3,7 +3,7 @@ fgivenx:
 ===================
 :fgivenx:  Functional Posterior Plotter 
 :Author: Will Handley
-:Version: 2.1.2
+:Version: 2.1.3
 :Homepage: https://github.com/williamjameshandley/fgivenx
 :Documentation: http://fgivenx.readthedocs.io/
 
