+++ .environment.yml
@@ -5,13 +5,14 @@ dependencies:
     - python=3.5
     - python=3.4
     - python=2.7
-    - matplotlib
-    - scipy
-    - numpy
-    - pypandoc
-    - numpydoc
-    - getdist
-    - joblib
-    - tqdm
-    - pytest
-    - pytest-cov
+    - matplotlib 
+    - scipy      
+    - numpy      
+    - sphinx     
+    - numpydoc   
+    - pytest     
+    - pytest-cov 
+    - joblib     
+    - tqdm       
+    - getdist    
+    - codecov    
