### 2018-08-04 d1fee81
Updated README description
