### 2018-08-04 135fa02
Updated the paper bibliography
