+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{inflation2015,
     author = {{Planck Collaboration}},
     title = "{Planck 2015 results. XX. Constraints on inflation}",
-    journal = {\aap},
+    journal = {A\&A},
     archivePrefix = "arXiv",
     eprint = {1502.02114},
     keywords = {cosmic background radiation, cosmology: theory, early Universe, inflation},
@@ -48,7 +48,7 @@
     author = {{Hee}, S. and {Handley}, W.~J. and {Hobson}, M.~P. and {Lasenby}, A.~N.
     },
     title = "{Bayesian model selection without evidences: application to the dark energy equation-of-state}",
-    journal = {\mnras},
+    journal = {MNRAS},
     archivePrefix = "arXiv",
     eprint = {1506.09024},
     keywords = {equation of state, methods: data analysis, methods: statistical, cosmological parameters, dark energy},
@@ -65,7 +65,7 @@
     author = {{Hee}, S. and {V{\'a}zquez}, J.~A. and {Handley}, W.~J. and 
         {Hobson}, M.~P. and {Lasenby}, A.~N.},
     title = "{Constraining the dark energy equation of state using Bayes theorem and the Kullback-Leibler divergence}",
-    journal = {\mnras},
+    journal = {MNRAS},
     archivePrefix = "arXiv",
     eprint = {1607.00270},
     keywords = {equation of state, methods: data analysis, methods: statistical, cosmological parameters, dark energy},
@@ -94,7 +94,8 @@
 }
 
 @ARTICLE{Higson2018,
-    author = {{Higson}, E. and {Handley}, W. and {Hobson}, M. and {Lasenby}, A.  },
+    author = {{Higson}, E. and {Handley}, W. and {Hobson}, M. and {Lasenby}, A.
+    },
     title = "{Diagnostic Tests for Nested Sampling Calculations}",
     journal = {ArXiv e-prints},
     archivePrefix = "arXiv",
@@ -119,8 +120,10 @@
 }
 
 @unpublished{Higson2018b,
-    author = {{Higson}, E. and {Handley}, W. and {Hobson}, M. and {Lasenby}, A.},
+    author = {{Higson}, E. and {Handley}, W. and {Hobson}, M. and {Lasenby}, A.
+    },
     title = "Bayesian Sparse Reconstruction: a brute-force approach to astronomical imaging and machine learning",
+    Year = {2018},
     note = "In preparation",
 }
 
