### 2018-08-04 9618683
Added further python versions to build matrix
