+++ .travis.yml
@@ -1,39 +1,21 @@
 language: python
-jobs:
-    include:
-        - stage: minimal
-          python:
-              - "2.7"
-              - "3.4"
-              - "3.5"
-              - "3.6"
-          install:
-              - pip install -r minimal_requirements.txt
-          before_script:
-              - "export DISPLAY=:99.0"
-              - "sh -e /etc/init.d/xvfb start"
-              - sleep 3
-          script: 
-              - pytest --cov=fgivenx
-          after_success: 
-              - codecov
-
-        - stage: full
-          install:
-              - pip install -r requirements.txt
-          python:
-              - "2.7"
-              - "3.4"
-              - "3.5"
-              - "3.6"
-          install:
-              - pip install -r minimal_requirements.txt
-          before_script:
-              - "export DISPLAY=:99.0"
-              - "sh -e /etc/init.d/xvfb start"
-              - sleep 3
-          script: 
-              - pytest --cov=fgivenx
-          after_success: 
-              - codecov
-
+python:
+    - "2.6"
+    - "2.7"
+    - "3.3"
+    - "3.4"
+    - "3.5"
+    - "3.6"
+env:
+    - REQUIREMENTS=minimal_requirements.txt
+    - REQUIREMENTS=requirements.txt
+install:
+    - pip install -r $REQUIREMENTS
+before_script:
+    - "export DISPLAY=:99.0"
+    - "sh -e /etc/init.d/xvfb start"
+    - sleep 3
+script: 
+    - pytest --cov=fgivenx
+after_success: 
+    - codecov
