### 2018-08-04 6bd3fcf
readthedocs build fixed
