+++ .environment.yml
@@ -2,17 +2,14 @@ name: fgivenx-read-the-docs-environment
 
 dependencies:
     - python=3.6
-    - python=3.5
-    - python=3.4
     - python=2.7
-    - matplotlib 
-    - scipy      
-    - numpy      
-    - sphinx     
-    - numpydoc   
-    - pytest     
-    - pytest-cov 
-    - joblib     
-    - tqdm       
-    - getdist    
-    - codecov    
+    - matplotlib
+    - tqdm
+    - pytest
+    - scipy
+    - numpy
+    - joblib
+    - pypandoc
+    - numpydoc
+    - codecov
+    - pytest-cov
