+++ readthedocs.yml
@@ -1,2 +1,5 @@
-conda:
-    file: .environment.yml
+python:
+    version: 3.6
+    pip_install: true
+    extra_requirements:
+        - docs
