+++ readthedocs.yml
@@ -1,5 +1,4 @@
 python:
-    version: 3.6
     pip_install: true
     extra_requirements:
         - docs
