+++ setup.py
@@ -23,21 +23,6 @@ def get_version(short=False):
                     return ver
 
 
-class PyTest(Command):
-    user_options = []
-    def initialize_options(self):
-        pass
-
-    def finalize_options(self):
-        pass
-
-    def run(self):
-        import subprocess
-        import sys
-        errno = subprocess.call(['pytest', '--cov=fgivenx'])
-        raise SystemExit(errno)
-
-
 setup(name='fgivenx',
       version=get_version(),
       description='fgivenx: Functional Posterior Plotter',
@@ -45,11 +30,15 @@ setup(name='fgivenx',
       author='Will Handley',
       author_email='wh260@cam.ac.uk',
       url='https://github.com/williamjameshandley/fgivenx',
-      packages=['fgivenx', 'fgivenx.tests'],
+      packages=['fgivenx'],
       install_requires=['matplotlib', 'numpy', 'scipy'],
-      extra_requires=['joblib', 'tqdm','getdist'],
-      extras_require={'docs': ['sphinx', 'sphinx_rtd_theme', 'numpydoc']},
-      cmdclass = {'test': PyTest},
+      setup_requires=['pytest-runner'],
+      extras_require={
+          'docs': ['sphinx', 'sphinx_rtd_theme', 'numpydoc'],
+          'parallel': ['joblib'],
+          'progress_bar': ['tqdm'],
+          'getdist_chains': ['getdist']
+          },
       tests_require=['pytest', 'pytest-cov', 'coverage'],
       license='MIT',
       classifiers=[
