+++ docs/source/fgivenx.tests.rst
@@ -1,54 +1,0 @@
-fgivenx.tests package
-=====================
-
-Submodules
-----------
-
-fgivenx.tests.test\_dkl module
-------------------------------
-
-.. automodule:: fgivenx.tests.test_dkl
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-fgivenx.tests.test\_io module
------------------------------
-
-.. automodule:: fgivenx.tests.test_io
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-fgivenx.tests.test\_mass module
--------------------------------
-
-.. automodule:: fgivenx.tests.test_mass
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-fgivenx.tests.test\_parallel module
------------------------------------
-
-.. automodule:: fgivenx.tests.test_parallel
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-fgivenx.tests.test\_samples module
-----------------------------------
-
-.. automodule:: fgivenx.tests.test_samples
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: fgivenx.tests
-    :members:
-    :undoc-members:
-    :show-inheritance:
