+++ setup.py
@@ -30,7 +30,7 @@ setup(name='fgivenx',
       author='Will Handley',
       author_email='wh260@cam.ac.uk',
       url='https://github.com/williamjameshandley/fgivenx',
-      packages=['fgivenx'],
+      packages=['fgivenx', 'fgivenx.test'],
       install_requires=['matplotlib', 'numpy', 'scipy'],
       setup_requires=['pytest-runner'],
       extras_require={
