### 2018-08-04 4a08258
Create CODE_OF_CONDUCT.md
