### 2018-08-04 100b0c6
Update PULL_REQUEST_TEMPLATE.md
