### 2018-08-04 #849 @williamjameshandley
I've now updated both pip and readthedocs so that they mirror the README.rst file in the GitHub repo, and added the additional information that was requested to the README. I also expanded the test suite so that it tests with and without the optional dependencies
