### 2018-08-05 1912d11
Reduced requirements on pytest, factoring them into .travis.yml
