+++ minimal_requirements.txt
@@ -1,6 +1,3 @@
 matplotlib
 scipy
 numpy
-pytest
-pytest-cov
-codecov
