### 2018-08-05 e8bd76f
Added a warning if parallelisation is not installed
