### 2018-08-05 1986a06
Fixed broken parallelisation for minimal dependency
