### 2018-08-09 #849 @FaustinCarter
Round 2
======

OK, I have some new comments! I actually tried it out on some data and made a pretty sweet plot of some residuals. I had to wrap my function to get the order of the args correct for your code (and to make it plot residuals rather than the function), but it worked just fine.

<img src="https://user-images.githubusercontent.com/10946756/43928316-0d9e8d04-9bf6-11e8-8e64-be6f252fe0a6.png" width=300/><img src="https://user-images.githubusercontent.com/10946756/43928322-10c136b2-9bf6-11e8-9c3a-3326eb5c4616.png" width=300/>

I also reinstalled the latest version in a fresh environment and all the tests ran just fine (although installing matplolib via pip in a conda environment on OSX breaks everything...). OK, now onto comments. Once these are addressed I can check off the last few boxes above and recommend you for publication.

General
--------
 * The version on PyPI is 2.1.8. The latest release on github is 2.0.0, which means the Zenodo DOI is for v. 2.0.0. I think if you generate a new release of 2.1.8, then the Zenodo hook should automatically update you with a new DOI for 2.1.8. Then I can check the version box on the checklist above.

 * Citing getdist: I ran into the same problem when I was trying to cite them. I eventually settled with the following citation: `Lewis, Antony. 2015. “Getdist Github Repository.” https://github.com/cmbant/getdist`

 * I think that the file outlining contributions could be added to the end of the readme. That's a more natural place for it, and it encourages people to get involved right from the start!

 * In general I strongly feel that` __init__.py` should be reserved for initializing things. This is perhaps my interpretation of what being Pythonic should mean, but in my view a better (and more consistent) approach would be to put your top-level functions in a file called fgivenx.py, and then use the `__init__.py` file to import those into the package namespace (`from .fgivenx import *`, or just defining them in `__all__`). Now you have access to those from the top level, but your users can easily locate them in your package (instead of hunting around for ages until finally thinking to look in the `__init__.py` file). The Hitchhiker's Guide to Python agrees on this point (https://docs.python-guide.org/writing/structure/), but if you feel strongly about keeping it the way it is, it's your package! I'm not going to make that a requirement of the review.

 * You might consider also generating a conda-forge recipe for your package in the future. Makes it easy for conda users to manage the dependencies. This is just a weak suggestion for an eventual future and doesn't affect this review.

Documentation
---------------
 * In the introduction of the docs, you say the driving routine is `compute_contours`, but your example does not explicitly call that function. In fact, I'm not sure that funciton exists? Do you mean `compute_pmf`?

 * The docstring for `fgivenx.samples.compute_samples` refers to the docstring for `compute_contours`. This shoudl be fixed to reference the correct function.

 * The "Introduction" should be the first thing one sees upon going to the docs website, and then there can be a link to the API. Especially since the figure should directly follow the example code in the introduction, and now it is on the landing page, but doesn't show up after the example code in the introduction.

 * It would be good to start a changelog and include that in the documentation somewhere.

* The docstring for `compute_samples` does not mention parameter `logZ`. Also does not mention that if `logZ` (apparently an array) is passed, then the user must pass a list of functions to `f` instead of a single function.

* The docstring of `compute_pmf` could be more explicit that the return value is a tuple. I had to actually go look at the code to figure out what it was really returning.

* I suggest a comprehensive review of all the docstrings to make sure they accurately reflect the function signature (and return value) and describe all the parameters, args, and kwargs. Also, each parameter should be defined in the doc string rather than saying "go look at this other docstring". Referring to another docstring is fine for kwargs that will be passed through to some other function, but not for required parameters.

Plotting
--------
* For plotting routines, ax could be an optional argument. Then, if not supplied, it creates an axis, plots on it, and returns the axis (or maybe figure) back again. This is more of a suggestion than anything, but it makes the plotting a bit more friendly for matplotlib novices because then `fgivenx.plot.plot` just gives you a nice figure. I'm not going to hold the review up on this point.

 * It's not clear to me what the plots test is actually doing. Ideally it would generate an image and compare it to some pre-rendered image, but I don't see you loading pytest-mpl at all. In fact, the second test doesn't even use pytest at all I don't think.

Paper
------
This is much, much better now. Just a few additional comments.
 * You can replace "we" with "I" everywhere since you are the only author and you aren't the queen.
 * Paragraph 3: should be "... a probability distribution..."
 * Paragraph 3: ref to Fig. 2 should include reference to the appropriate panel.

Random
--------
I'm mystified by line 37 in `samples.py`. It looks like you are zipping two lists of different length together for the case where `logZ` is omitted. That should just return an iterator that is the length of the shortest list (which I think is length 1), but somehow it doesn't? I see where you set `f=[f]`, but not where you make that new list the same length as the samples dimension you are zipping against. How does this even work?
