+++ fgivenx/drivers.py
@@ -1,0 +1,236 @@
+r"""This module provides utilities for computing the grid for contours of a
+    function reconstruction plot.
+
+    Required ingredients:
+     * sampled posterior probability distribution :math:`P(\theta)`
+     * independent variable :math:`x`
+     * dependent variable :math:`y`
+     * functional form :math:`y = f(x;\theta)` parameterised by :math:`\theta`
+
+    Assuming that you have obtained samples of theta from an MCMC
+    process, we aim to compute the density:
+
+    .. math::
+
+        P(y|x) &= \int P(y=f(x;\theta)|x,\theta) P(\theta) d\theta \\
+                &= \int \delta(y-f(x;\theta)) P(\theta) d\theta
+
+    which gives our degree of knowledge for each :math:`y` value given an
+    :math:`x` value.
+
+    In fact, for a more representative plot, we are not actually
+    interested in the value of the probability density above, but in fact
+    require the "iso-probablity posterior mass"
+
+    .. math::
+
+        \mathrm{pmf}(y|x) = \int_{P(y'|x) < P(y|x)} P(y'|x) dy'
+
+    We thus need to compute this function on a rectangular grid of :math:`x`
+    and :math:`y`.
+"""
+import numpy
+import fgivenx.samples
+import fgivenx.mass
+import fgivenx.dkl
+from fgivenx._utils import _check_args, _normalise_weights,\
+                           _equally_weight_samples
+
+
+def compute_samples(f, x, samples, logZ=None, **kwargs):
+    r"""
+    Apply the function(s) :math:`f(x;\theta)` to the arrays defined in x and
+    samples.  Has options for weighting, trimming, cacheing and parallelising.
+
+    Additionally, if a list of log-evidences are passed, along with list of
+    functions, samples and optional weights it marginalises over the models
+    according to the evidences.
+
+    Parameters
+    ----------
+    f: function
+        function :math:`f(x;\theta)` with dependent variable :math:`x`,
+        parameterised by :math:`\theta`.
+
+    x: 1D array-like
+        x values to evaluate :math:`f(x;\theta)` at.
+
+    samples: 2D array-like
+        theta samples to evaluate :math:`f(x;\theta)` at.
+        shape = (nsamples, npars)
+
+    weights: 1D array-like, optional
+        sample weights, if desired. Should have length same as samples.shape[0]
+        default `numpy.ones_like(samples)`
+
+    ntrim: int, optional
+        Approximate number of samples to trim down to, if desired. Useful if
+        the posterior is dramatically oversampled
+        default None
+
+    cache: str, optional
+        File root for saving previous calculations for re-use
+        default None
+
+    parallel, tqdm_args:
+        see docstring for :func:`fgivenx.parallel.parallel_apply`
+
+    Returns
+    -------
+    2D numpy.array
+        Evaluate the function f at each x value and each theta.
+        Equivalent to `[[f(x_i,theta) for theta in samples] for x_i in x]`
+
+    """
+    weights = kwargs.pop('weights', None)
+    parallel = kwargs.pop('parallel', False)
+    ntrim = kwargs.pop('ntrim', None)
+    cache = kwargs.pop('cache', None)
+    tqdm_kwargs = kwargs.pop('tqdm_kwargs', {})
+    if kwargs:
+        raise TypeError('Unexpected **kwargs: %r' % kwargs)
+
[…]
