+++ fgivenx/__init__.py
@@ -1,1 +1,1 @@
-from fgivenx.drivers import compute_samples, compute_pmf, compute_dkl
+from fgivenx.drivers import plot_contours, plot_lines, plot_dkl, compute_samples, compute_pmf, compute_dkl
