### 2018-08-11 dbde1b2
Updated documentation
