+++ README.rst
@@ -40,7 +40,7 @@ If there is a function parameterised by theta ``f(x;theta)``, then this script
 will produce a contour plot of the conditional posterior ``P(f|x,D,M)`` in the
 ``(x,f)`` plane.
 
-The driving routine is ``fgivenx.compute_contours``, and `Example Usage`_ can
+The driving routines are ``fgivenx.plot_lines``, and `Example Usage`_ can
 be found by running ``help(fgivenx)``. It is compatible with getdist, and has a
 loading function provided by ``fgivenx.samples.samples_from_getdist_chains``.
 
@@ -151,97 +151,100 @@ Example Usage
 
 .. code:: python
 
-   import numpy
-   import matplotlib.pyplot as plt
-   from fgivenx import compute_samples, compute_pmf, compute_dkl
-   from fgivenx.plot import plot, plot_lines
+    import numpy
+    import matplotlib.pyplot as plt
+    from fgivenx import plot_contours, plot_lines, plot_dkl
 
 
-   # Model definitions
-   # =================
-   # Define a simple straight line function, parameters theta=(m,c)
-   def f(x, theta):
-       m, c = theta
-       return m * x + c
+    # Model definitions
+    # =================
+    # Define a simple straight line function, parameters theta=(m,c)
+    def f(x, theta):
+        m, c = theta
+        return m * x + c
 
 
-   numpy.random.seed(1)
+    numpy.random.seed(1)
 
-   # Posterior samples
-   nsamples = 1000
-   ms = numpy.random.normal(loc=-5, scale=1, size=nsamples)
-   cs = numpy.random.normal(loc=2, scale=1, size=nsamples)
-   samples = numpy.array([(m, c) for m, c in zip(ms, cs)]).copy()
+    # Posterior samples
+    nsamples = 1000
+    ms = numpy.random.normal(loc=-5, scale=1, size=nsamples)
+    cs = numpy.random.normal(loc=2, scale=1, size=nsamples)
+    samples = numpy.array([(m, c) for m, c in zip(ms, cs)]).copy()
 
-   # Prior samples
-   ms = numpy.random.normal(loc=0, scale=5, size=nsamples)
-   cs = numpy.random.normal(loc=0, scale=5, size=nsamples)
-   prior_samples = numpy.array([(m, c) for m, c in zip(ms, cs)]).copy()
+    # Prior samples
+    ms = numpy.random.normal(loc=0, scale=5, size=nsamples)
+    cs = numpy.random.normal(loc=0, scale=5, size=nsamples)
+    prior_samples = numpy.array([(m, c) for m, c in zip(ms, cs)]).copy()
 
-   # Computation
-   # ===========
-   # Examine the function over a range of x's
-   xmin, xmax = -2, 2
-   nx = 100
-   x = numpy.linspace(xmin, xmax, nx)
+    # Set the x range to plot on
+    xmin, xmax = -2, 2
+    nx = 100
+    x = numpy.linspace(xmin, xmax, nx)
 
-   # Set the cache
-   cache = 'cache/test'
-   prior_cache = cache + '_prior'
+    # Set the cache
+    cache = 'cache/test'
+    prior_cache = cache + '_prior'
 
-   # Compute function samples
-   fsamps = compute_samples(f, x, samples, cache=cache)
-   prior_fsamps = compute_samples(f, x, prior_samples, cache=prior_cache)
+    # Plotting
+    # ========
+    fig, axes = plt.subplots(2, 2)
 
-   # Compute dkls
[…]
