+++ paper/paper.bib
@@ -148,6 +148,14 @@
     note = {[Online; accessed <today>]}
 }
 
+@Misc{getdist,
+    author =    {Lewis, Anthony},
+    title =     {Getdist Github Repository},
+    year =      {2015},
+    url = "https://github.com/cmbant/getdist",
+    note = {[Online; accessed <today>]}
+}
+
 @article{corner,
     Author = {Daniel Foreman-Mackey},
     Doi = {10.21105/joss.00024},
