### 2018-08-11 9c00473
Fixed plotting coverage
