+++ fgivenx/test/test_z.py
@@ -1,7 +1,7 @@
 import numpy
 import matplotlib.pyplot as plt
-from fgivenx import compute_samples, compute_pmf, compute_dkl
-from fgivenx.plot import plot, plot_lines
+from fgivenx import plot_contours, plot_lines, plot_dkl
+from fgivenx.drivers import compute_samples, compute_pmf, compute_dkl
 import pytest
 
 
@@ -39,21 +39,21 @@ def test_full():
 
     # Compute function samples
     with pytest.raises(TypeError):
-        fsamps = compute_samples(f, x, samples, unknown_arg=None)
-    fsamps = compute_samples(f, x, samples, cache=cache)
-    prior_fsamps = compute_samples(f, x, prior_samples, cache=prior_cache)
+        compute_samples(f, x, samples, unknown_arg=None)
+    compute_samples(f, x, samples, cache=cache)
+    compute_samples(f, x, prior_samples, cache=prior_cache)
 
     # Compute dkls
     with pytest.raises(TypeError):
-        fsamps = compute_dkl(f, x, samples, prior_samples, unknown_arg=None)
-    dkls = compute_dkl(f, x, samples, prior_samples,
-                       cache=cache, parallel=True)
+        compute_dkl(f, x, samples, prior_samples, unknown_arg=None)
+    compute_dkl(f, x, samples, prior_samples,
+                cache=cache, parallel=True)
 
     # Compute probability mass function.
     with pytest.raises(TypeError):
-        fsamps = compute_pmf(f, x, samples, unknown_arg=None)
+        compute_pmf(f, x, samples, unknown_arg=None)
     with pytest.raises(ValueError):
-        fsamps = compute_pmf(f, x, samples, y=numpy.random.rand(2, 2))
+        compute_pmf(f, x, samples, y=numpy.random.rand(2, 2))
     y, pmf = compute_pmf(f, x, samples, cache=cache, parallel=True)
     y_prior, pmf_prior = compute_pmf(f, x, prior_samples,
                                      cache=prior_cache, parallel=True)
@@ -61,43 +61,42 @@ def test_full():
     # Plotting
     # ========
     fig, axes = plt.subplots(2, 2)
-    prior_color = 'b'
-    posterior_color = 'r'
 
     # Sample plot
     # -----------
     ax_samples = axes[0, 0]
     ax_samples.set_ylabel(r'$c$')
     ax_samples.set_xlabel(r'$m$')
-    ax_samples.plot(prior_samples.T[0], prior_samples.T[1],
-                    color=prior_color, marker='.', linestyle='')
-    ax_samples.plot(samples.T[0], samples.T[1],
-                    color=posterior_color, marker='.', linestyle='')
+    ax_samples.plot(prior_samples.T[0], prior_samples.T[1], 'b.')
+    ax_samples.plot(samples.T[0], samples.T[1], 'r.')
 
     # Line plot
     # ---------
     ax_lines = axes[0, 1]
     ax_lines.set_ylabel(r'$y = m x + c$')
     ax_lines.set_xlabel(r'$x$')
-    plot_lines(x, prior_fsamps, ax_lines, color=prior_color)
-    plot_lines(x, fsamps, ax_lines, color=posterior_color)
+    plot_lines(f, x, prior_samples, ax_lines, color='b', cache=prior_cache)
+    plot_lines(f, x, samples, ax_lines, color='r', cache=cache)
 
     # Predictive posterior plot
     # -------------------------
     ax_fgivenx = axes[1, 1]
[…]
