+++ fgivenx/parallel.py
@@ -2,29 +2,25 @@ import warnings
 try:
     from tqdm import tqdm, tqdm_notebook
 except ImportError:
-    def tqdm(x, **kwargs):
-        return x
+    def tqdm(x, **kwargs): return x
 
     def tqdm_notebook(x, **kwargs): return x
 
 
 try:
-    PARALLEL=True
+    PARALLEL = True
     from joblib import Parallel, delayed, cpu_count
 except ImportError:
-    PARALLEL=False
+    PARALLEL = False
+
     class Parallel(object):
-        def __init__(self, n_jobs=None):
-            pass
+        def __init__(self, n_jobs=None): pass
 
-        def __call__(self, x):
-            return list(x)
+        def __call__(self, x): return list(x)
 
-    def delayed(x):
-        return x
+    def delayed(x): return x
 
-    def cpu_count():
-        return 1
+    def cpu_count(): return 1
 
 
 def parallel_apply(f, array, **kwargs):
@@ -71,7 +67,7 @@ def parallel_apply(f, array, **kwargs):
         raise TypeError('Unexpected **kwargs: %r' % kwargs)
     try:
         # If running in a jupyter notebook then use tqdm_notebook.
-        if get_ipython().has_trait('kernel'): progress = tqdm_notebook
+        progress = tqdm_notebook if get_ipython().has_trait('kernel') else tqdm
     except (NameError, AssertionError):
         # Otherwise use regular tqdm progress bar
         progress = tqdm
@@ -89,7 +85,8 @@ def parallel_apply(f, array, **kwargs):
         raise ValueError("parallel keyword must be an integer or bool")
 
     if parallel and not PARALLEL:
-        warnings.warn("You need to install the package joblib if you want to use parallelisation")
+        warnings.warn("You need to install the package joblib"
+                      "if you want to use parallelisation")
 
     return Parallel(n_jobs=parallel)(delayed(f)(*(precurry + (x,) + postcurry))
-                                   for x in progress(array, **tqdm_kwargs))
+                                     for x in progress(array, **tqdm_kwargs))
