### 2018-08-11 9dbe886
Added plot image test
