+++ .travis.yml
@@ -11,9 +11,7 @@ install:
     - pip install -r $REQUIREMENTS
     - pip install pytest pytest-cov codecov
 before_script:
-    - "export DISPLAY=:99.0"
-    - "sh -e /etc/init.d/xvfb start"
-    - sleep 3
+    - "export MPLBACKEND=Agg"
 script: 
     - pytest --cov=fgivenx
 after_success: 
