+++ docs/source/index.rst
@@ -1,23 +1,3 @@
-fgivenx
-=======
+.. title:: Introduction
 
-fgivenx is a Python package for plotting predictive posteriors of functions
-
-.. toctree::
-   :maxdepth: 1
-
-   Introduction <intro>
-
-|image0|
-
-Modules:
-
-.. toctree::
-   :maxdepth: 2
-
-   fgivenx
-
-* :ref:`genindex`
-* :ref:`search`
-
-.. |image0| image:: https://raw.github.com/williamjameshandley/fgivenx/master/plot.png
+.. include:: ../../README.rst
