+++ docs/source/fgivenx.test.rst
@@ -1,0 +1,78 @@
+fgivenx.test package
+====================
+
+Submodules
+----------
+
+fgivenx.test.test\_\_utils module
+---------------------------------
+
+.. automodule:: fgivenx.test.test__utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fgivenx.test.test\_dkl module
+-----------------------------
+
+.. automodule:: fgivenx.test.test_dkl
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fgivenx.test.test\_drivers module
+---------------------------------
+
+.. automodule:: fgivenx.test.test_drivers
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fgivenx.test.test\_io module
+----------------------------
+
+.. automodule:: fgivenx.test.test_io
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fgivenx.test.test\_mass module
+------------------------------
+
+.. automodule:: fgivenx.test.test_mass
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fgivenx.test.test\_parallel module
+----------------------------------
+
+.. automodule:: fgivenx.test.test_parallel
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fgivenx.test.test\_plot module
+------------------------------
+
+.. automodule:: fgivenx.test.test_plot
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fgivenx.test.test\_samples module
+---------------------------------
+
+.. automodule:: fgivenx.test.test_samples
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: fgivenx.test
+    :members:
+    :undoc-members:
+    :show-inheritance:
