### 2018-08-11 b9bcde2
Updated samples to leverage getdist properly
