+++ docs/source/index.rst
@@ -1,3 +1,17 @@
-.. title:: Introduction
+fgivenx
+=======
 
-.. include:: ../../README.rst
+.. toctree::
+   :maxdepth: 1
+   :caption: Contents
+
+   Introduction <intro>
+   fgivenx
+
+
+Indices and tables
+------------------
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
