### 2018-08-11 cc82749
Added alpha to the line plots
