### 2018-08-11 751c93c
Added figures further down the readme
