+++ docs/source/fgivenx.rst
@@ -1,28 +1,21 @@
 fgivenx package
 ===============
 
-Subpackages
------------
-
-.. toctree::
-
-    fgivenx.test
-
 Submodules
 ----------
-
-fgivenx.dkl module
-------------------
-
-.. automodule:: fgivenx.dkl
-    :members:
-    :undoc-members:
-    :show-inheritance:
 
 fgivenx.drivers module
 ----------------------
 
 .. automodule:: fgivenx.drivers
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fgivenx.dkl module
+------------------
+
+.. automodule:: fgivenx.dkl
     :members:
     :undoc-members:
     :show-inheritance:
