### 2018-08-12 aecadda
Further updates to docstrings
