### 2018-08-12 5e95801
Further updates to the paper
