+++ paper/paper.bib
@@ -141,9 +141,9 @@
 }
 
 @Misc{scipy,
-    author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+    author =    {{Jones}, Eric  and {Oliphant}, Travis and {Peterson}, Pearu},
     title =     {{SciPy}: Open source scientific tools for {Python}},
-    year =      {2001--},
+    year =      {2001},
     url = "http://www.scipy.org/",
     note = {[Online; accessed <today>]}
 }
