### 2018-08-12 #849 @williamjameshandley
Nice plots!
