### 2018-08-12 #849 @williamjameshandley
> General
> ----------
> * The version on PyPI is 2.1.8. The latest release on github is 2.0.0, which means the Zenodo DOI is for v. 2.0.0. I think if you generate a new release of 2.1.8, then the Zenodo hook should automatically update you with a new DOI for 2.1.8. Then I can check the version box on the checklist above.

I've updated the new version to 2.1.9, and created a release so that Zenodo is up-to-date.

> * Citing getdist: I ran into the same problem when I was trying to cite them. I eventually settled with the following citation: Lewis, Antony. 2015. “Getdist Github Repository.” https://github.com/cmbant/getdist
> * I think that the file outlining contributions could be added to the end of the readme. That's a more natural place for it, and it encourages people to get involved right from the start!

Both of these are now done.

> * In general I strongly feel that__init__.py should be reserved for initializing things...

I've done this, by moving the relevant code to `fgivenx/drivers.py` and importing it in `__init__.py`. I couldn't get `fgivenx/fgivenx.py` to work for python2 in a clean manner.

> * You might consider also generating a conda-forge recipe for your package in the future. Makes it easy for conda users to manage the dependencies. This is just a weak suggestion for an eventual future and doesn't affect this review.

I don't really have any experience of using conda, but know that a lot of people do. One way to make sure that this does get done in the future would be to post an 'issue' to the GitHub -- that way I'll be reminded/encouraged to get round to this at some point in the future.
