### 2018-08-12 48a1368
Added changelog
