### 2018-08-12 #849 @williamjameshandley
>Documentation
>----------------------
>* In the introduction of the docs, you say the driving routine is compute_contours, but your example does not explicitly call that function. In fact, I'm not sure that funciton exists? Do you mean compute_pmf?
>* The docstring for fgivenx.samples.compute_samples refers to the docstring for compute_contours. This shoudl be fixed to reference the correct function.

I've now introduced three new driver functions `fgivenx.plot_contours`, `fgivenx.plot_lines`, and `fgivenx.plot_dkl`, and changed all references to those

>* The "Introduction" should be the first thing one sees upon going to the docs website, and then there can be a link to the API. Especially since the figure should directly follow the example code in the introduction, and now it is on the landing page, but doesn't show up after the example code in the introduction.
>* It would be good to start a changelog and include that in the documentation somewhere.

Done

>* The docstring for compute_samples does not mention parameter logZ. Also does not mention that if logZ (apparently an array) is passed, then the user must pass a list of functions to f instead of a single function.

I think I've clarified this.

>* The docstring of compute_pmf could be more explicit that the return value is a tuple. I had to actually go look at the code to figure out what it was really returning.

Currently the return section of the docstring has a list of the two arguments. What do you think would make this clearer?

>* I suggest a comprehensive review of all the docstrings to make sure they accurately reflect the function signature (and return value) and describe all the parameters, args, and kwargs. Also, each parameter should be defined in the doc string rather than saying "go look at this other docstring". Referring to another docstring is fine for kwargs that will be passed through to some other function, but not for required parameters.

Done.
