### 2018-08-12 #849 @williamjameshandley
>Plotting
>-----------
> * For plotting routines, ax could be an optional argument. Then, if not supplied, it creates an axis, plots on it, and returns the axis (or maybe figure) back again. This is more of a suggestion than anything, but it makes the plotting a bit more friendly for matplotlib novices because then fgivenx.plot.plot just gives you a nice figure. I'm not going to hold the review up on this point.

I think this a great idea. I've leveraged plt.gca() to get this behaviour. If ax is not passed, then it just plots on the last axis, or creates one if there isn't one, consistent with the way that `matplotlib.pyplot.plot` works.

> * It's not clear to me what the plots test is actually doing. Ideally it would generate an image and compare it to some pre-rendered image, but I don't see you loading pytest-mpl at all. In fact, the second test doesn't even use pytest at all I don't think.

I've added a matplotlib test that checks that it produces the correct image to this test.
