### 2018-08-12 #849 @williamjameshandley
> Random
> ------------
> * I'm mystified by line 37 in samples.py. It looks like you are zipping two lists of different length together for the case where logZ is omitted. That should just return an iterator that is the length of the shortest list (which I think is length 1), but somehow it doesn't? I see where you set f=[f], but not where you make that new list the same length as the samples dimension you are zipping against. How does this even work?

This function assumes that it's given a list of functions and a list of samples, i.e. it's been already pre-processed to be evidence-weighted as in `fgivenx.drivers.compute_pmf`. I've updated the docstring to reflect this.
