### 2018-08-12 #849 @williamjameshandley
Many thanks for all these comments. I think this is really good process for improving distributed code.
