### 2018-08-14 #849 @FaustinCarter
Round 3
======

This is all looking great. There are just a couple lingering things I see in the docs:

 * The plot images aren't loading at Read the Docs
 * There is a ghost section in the API table of contents since you moved your main code out of the `__init__.py` file.

and with the tests:

 * You should add `pytest-mpl` as a testing dependency as tests will immediately fail without it.
 * For some reason the test image isn't installing from PyPI so the image comparison test fails; the entire baseline images directory is missing. The other tests all passed with the latest version installed via pip from PyPI.

I think once those are fixed and a new release is created to update the Zenodo DOI then I'll green light you.
