### 2018-08-15 f9b3963
Updates to documentation
