+++ paper/paper.bib
@@ -110,13 +110,12 @@
 
 @article{zenodo,
     Author = {Handley, W},
-    Date-Modified = {2018-07-12 11:21:25 +0000},
-    doi = {10.5281/zenodo.1310715},
-    Month = {Jul},
+    doi = {10.5281/zenodo.1344097},
+    Month = {Aug},
     Publisher = {Zenodo},
-    Title = {fgivenx: v2.0},
+    Title = {fgivenx: v2.1.10},
     Year = {2018},
-    Bdsk-Url-1 = {http://dx.doi.org/10.5281/zenodo.1310715}
+    Bdsk-Url-1 = {http://dx.doi.org/10.5281/zenodo.1344097}
 }
 
 @unpublished{Higson2018b,
