+++ MANIFEST.in
@@ -1,1 +1,2 @@
 include README.rst
+recursive-include fgivenx/test/baseline_images/ *.pdf
