+++ paper/paper.bib
@@ -110,12 +110,12 @@
 
 @article{zenodo,
     Author = {Handley, W},
-    doi = {10.5281/zenodo.1344097},
+    doi = {10.5281/zenodo.1345486},
     Month = {Aug},
     Publisher = {Zenodo},
-    Title = {fgivenx: v2.1.10},
+    Title = {fgivenx: v2.1.11},
     Year = {2018},
-    Bdsk-Url-1 = {http://dx.doi.org/10.5281/zenodo.1344097}
+    url = {http://dx.doi.org/10.5281/zenodo.1345486}
 }
 
 @unpublished{Higson2018b,
