### 2018-08-15 #849 @williamjameshandley
> * The plot images aren't loading at Read the Docs

Links fixed.

> * There is a ghost section in the API table of contents since you moved your main code out of the __init__.py file.

Added a section here with example usage and showing the global imports.

> * You should add pytest-mpl as a testing dependency as tests will immediately fail without it.

I don't actually use `pytest-mpl`, only the ones that come already packaged as `matplotlib.testing.decorators.image_comparison`, so I don't think you should need this (and [travis](https://travis-ci.org/williamjameshandley/fgivenx) seems to agree)

> * For some reason the test image isn't installing from PyPI so the image comparison test fails; the entire baseline images directory is missing. The other tests all passed with the latest version installed via pip from PyPI.

I've added the image comparison documents to the `Manifest.in`, so this should now work from PyPI

> I think once those are fixed and a new release is created to update the Zenodo DOI then I'll green light you.

I've also added a new release as [v2.1.11](https://doi.org/10.5281/zenodo.1345486), an explicit reference to the version number in the paper.md, and updated paper.bib

You may need to refresh caches to see the updates to the documentation (I did).
