### 2018-08-17 #849 @FaustinCarter
Round 4
======

Ok, I made a fresh environment with just: `python, matplotlib, scipy, numpy, pytest`. I then did `pip install fgivenx` and got version 2.1.11. I then ran the tests as: `pytest fgivenx`. I immediately get an `@imagecomparison` decorator error. As far as I can tell this is because `matplotlib` tests require at a minimum the `nose` package. I then installed `nose` and reran the tests. Everything went smoothly but the image compare test failed because the image with which to compare against is still not getting added to the pip installed directory.

I'm not sure I should hold you up on this point though because my package that I published via JOSS also doesn't actually load any of the tests via pip install. You have to get the code off github to run them. After this review is over I plan to fix that!

I think at a minimum if you can add either `nose` or `pytest-mpl` to the testing dependencies, and then make a note in the docs/readme about how to manually download the comparison image, then I'll call it good!
