### 2018-08-18 e77a7d2
updated test images and added pytest to tests in setup.py
