### 2018-08-18 d958af1
Further bug fixes to image comparison
