+++ setup.py
@@ -40,6 +40,7 @@ setup(name='fgivenx',
           'getdist_chains': ['getdist']
           },
       tests_require=['pytest', 'pytest-mpl'],
+      include_package_data=True,
       license='MIT',
       classifiers=[
                    'Development Status :: 5 - Production/Stable',
