### 2018-08-18 #849 @williamjameshandley
OK. I think I've sorted this out now (and updated the documentation)

The following pip commands in a fresh virtual environment work for me:
```bash
python -m venv joss
source joss/bin/activate

pip install fgivenx
pip install pytest pytest-mpl
export MPLBACKEND=Agg
pytest joss/lib/python3.7/site-packages/fgivenx
```

I can't unfortunately update the zenodo link in the paper since there is a 'major incident' currently occurring.
