### 2018-08-22 #849 @zhampel
@williamjameshandley Nice work with fgivenx. I have a few comments which I'll put in a couple of posts for you. I'll put up the comments regarding the paper and documentation now, and I'll have a couple more regarding the code itself either later today or early tomorrow.
