### 2018-08-22 #849 @zhampel
Paper
===
I think the paper could use a little cleanup. I found a few minor mistakes and would like to add some additional comments regarding the text as well. Take them as suggestions, but I feel like they'd make the paper smoother.

- Title after 'fgivenx:' should be capitalized

Paragraph 1
------------
- I find 'scientific models' not particularly specific or descriptive. Numerical maybe? Or change Scientists to Researchers.

Paragraph 2
------------
- Plots such as... I think 'plots' is a too colloquial. Find a better way to say this like 'Countour
- No dashes in Markov Chain Monte Carlo. But after writing it out, add '(MCMC)' for clarity.
- I think 'available freely online.' is better, where the word online is the link. Also add a period at the end of that sentence ;)

Paragraph 3
------------
- Generally, ensure that there’s no confusion about what probability we’re talking about. For example, one could in principle compare probabilities of a linear to quadratic functions, but that’s not what fgivenx does.
- Lines are straight, so I would change them all to line or linear as descriptors:
straight line - > line, straight line relationship -> linear relationship
- If one has a probability distribution -> Given probability distributions
- would also be uncertain -> characterized by some uncertainty.
Is the linear relationship uncertain? I think you should clarify it's the probability of assuming a linear function to predict y given x.
- The appropriate plot... -> This is depicted as P(y|x) in the bottom right panel of Figure 2.

Paragraph 4
------------
- Capitalize Python. This goes for all other mentions of Python as well
- I think 'fgivenx is a package for showing the relationships as depicted in Figure 2…` sounds a bit smoother
- is a little over-simplified -> provides a simple illustration
- Move the citation to the end of the phrase: 'in recent Planck studies… perturbations (citations),'

Paragraph 5
------------
- Capitalize Python.

Figure 1
--------
- may have noted -> may note
- Explain what the contours are. Confidence intervals with some quoted %?

Figure 2
--------
- caption has two periods
- Something change like this: An example of plots -> Example output of fgivenx provided some input data and linear test function.
