### 2018-08-22 #849 @zhampel
Documentation
===========
I appreciate the completeness that you put into the docstrings. I have a few comments.

Introduction
------------
In the 2nd paragraph of the Description
- These packages -> This package?

API
---
- no quotes necessary around iso-probability posterior mass
- Is logZ relative evidences w.r.t. models? I.e. is it the ratio P/Q? Or is it some other type of function prior?
Is this Q(\theta)? Z is not used in any of the latex equations.
And maybe make it a little more explicit that the user need not normalize it.
I know you say relative, but just make it super clear so there’s no possible confusion.
Also please make distinction between evidences comparing models and parameter priors, just to be safe.
- I suppose just ensure that any and all python variables that are part of equations are clearly and properly annotated in the docstrings.
- theta samples -> can you make it latex or mathjax, and also say 'parameters \theta' so it's clear
