### 2018-08-22 #849 @zhampel
At the moment I have only one comment in my notes regarding the code:
-I really like the kwargs.pop functionality you put it. So smooth :)

For now, it's really just ensuring the docstrings are uber clear, and I'll finish up my code review.
