### 2018-08-22 386f544
Updated paper after comments from second reviewer
