+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'fgivenx: a python package for functional posterior plotting'
+title: 'fgivenx: A Python package for functional posterior plotting'
 tags:
   - Python
   - Statistics
@@ -24,57 +24,59 @@ bibliography: paper.bib
 
 # Summary
 
-Scientists are often concerned with numerical values of parameters in
-scientific models. Our knowledge of such things can be quantified and presented
+Researchers are often concerned with numerical values of parameters in
+numerical models. Our knowledge of such things can be quantified and presented
 using probability distributions as demonstrated in Figure 1.
 
 ![The age and size of the universe, as measured using Planck 2018 data.
-(non-Astro)Physicists may have noted that 14 Gigaparsecs is roughly 46 billion
-light years. The fact that the observable universe is roughly three times
-larger in light years in comparison with its age is explained by the expansion
-of space over cosmic history.](planck.png) 
+(non-Astro)Physicists may note that 14 Gigaparsecs is roughly 46 billion light
+years. The fact that the observable universe is roughly three times larger in
+light years in comparison with its age is explained by the expansion of space
+over cosmic history. Contours indicate 67% and 95% marginalised iso-probability
+credibility regions](planck.png) 
 
-Plots such as Figure 1 can be created using two-dimensional kernel density
-estimation using packages such as
+Contour plots such as Figure 1 can be created using two-dimensional kernel
+density estimation using packages such as
 [scipy](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.gaussian_kde.html)
 [@scipy], [getdist](http://getdist.readthedocs.io/en/latest/intro.html)
 [@getdist], [corner](https://corner.readthedocs.io/en/latest/) [@corner] and
 [pygtc](https://pygtc.readthedocs.io/en/latest/) [@pygtc], where the samples
 provided as inputs to such programs are typically created by a
-Markov-Chain-Monte-Carlo analysis. For further information on MCMC and Bayesian
-analysis in general, "Information Theory, Inference and Learning Algorithms" is
-highly recommended [@mackay], which is available for free as an online
-[pdf](http://www.inference.org.uk/itprnn/book.html)
+Markov Chain Monte Carlo (MCMC) analysis. For further information on MCMC and
+Bayesian analysis in general, "Information Theory, Inference and Learning
+Algorithms" is highly recommended [@mackay], which is available freely
+[online](http://www.inference.org.uk/itprnn/book.html).
 
 As well as quantifying the uncertainty of real-valued parameters, scientists
-may also be interested in producing a probability distribution for a function
-``f(x)``. Take as a universally-relatable case the equation of a straight line
-``y = m*x + c``. If one has a probability distribution for the gradient ``m``
-and intercept ``c``, then the ability to predict ``y`` knowing ``x``
-using the straight line relationship would also be uncertain. The appropriate
-plot to make is that of ``P(y|x)`` in the ``(x,y)`` plane, as demonstrated in
-Figure 2's bottom right panel.
+may also be interested in producing a probability distribution for the
+predictive posterior of a function ``f(x)``. Take as a universally-relatable
+case the equation of a  line ``y = m*x + c``. Given posterior probability
+distributions for the gradient ``m`` and intercept ``c``, then the ability to
+predict ``y`` knowing ``x`` given their linear relationship would also be
+characterized by some uncertainty. This is depicted as ``P(y|x)`` in the bottom
+right panel of Figure 2.
 
-![An example of plots produced by fgivenx..
+![Example output of fgivenx provided some prior and posterior samples and a
+linear test function.
 Top-left: underlying parameter covariances between ``m`` and ``c`` for
[…]
