### 2018-08-22 c77de78
Docstring comment updates
