+++ fgivenx/drivers.py
@@ -7,7 +7,7 @@ r"""This module provides utilities for computing the grid for contours of a
      * dependent variable :math:`y`
      * functional form :math:`y = f(x;\theta)` parameterised by :math:`\theta`
 
-    Assuming that you have obtained samples of theta from an MCMC
+    Assuming that you have obtained samples of :math:`\theta` from an MCMC
     process, we aim to compute the density:
 
     .. math::
@@ -15,8 +15,8 @@ r"""This module provides utilities for computing the grid for contours of a
         P(y|x) &= \int P(y=f(x;\theta)|x,\theta) P(\theta) d\theta \\
                &= \int \delta(y-f(x;\theta)) P(\theta) d\theta
 
-    which gives our degree of knowledge for each :math:`y` value given an
-    :math:`x` value.
+    which gives our degree of knowledge for each :math:`y=f(x;\theta)` value
+    given an :math:`x` value.
 
     In fact, for a more representative plot, we are not actually
     interested in the value of the probability density above, but in fact
@@ -41,7 +41,7 @@ from fgivenx._utils import _check_args, _normalise_weights,\
 
 def plot_contours(f, x, samples, ax=None, **kwargs):
     r"""
-    Plot the probability mass function given x at a range of y values
+    Plot the probability mass function given `x` at a range of :math:`y` values
     for :math:`y = f(x|\theta)`
 
     :math:`P(y|x) = \int P(y=f(x;\theta)|x,\theta) P(\theta) d\theta`
@@ -50,7 +50,7 @@ def plot_contours(f, x, samples, ax=None, **kwargs):
 
     Additionally, if a list of log-evidences are passed, along with list of
     functions, and list of samples, this function plots the probability mass
-    function of all models marginalised according to the evidences.
+    function for all models marginalised according to the evidences.
 
     Parameters
     ----------
@@ -59,10 +59,10 @@ def plot_contours(f, x, samples, ax=None, **kwargs):
         dependent variable :math:`x`, parameterised by :math:`\theta`.
 
     x: 1D array-like
-        x values to evaluate :math:`f(x;\theta)` at.
+        `x` values to evaluate :math:`f(x;\theta)` at.
 
     samples: 2D array-like
-        theta samples (or list of theta samples) to evaluate
+        :math:`\theta` samples (or list of :math:`\theta` samples) to evaluate
         :math:`f(x;\theta)` at.
         `shape = (nsamples, npars)`
 
@@ -72,27 +72,27 @@ def plot_contours(f, x, samples, ax=None, **kwargs):
         get the last axis used, or create a new one.
 
     logZ: 1D array-like, optional
-        relative evidences of each model if multiple models are passed.
-        Should be same length as the list f
-        default: `numpy.ones_like(f)`
+        log-evidences of each model if multiple models are passed.
+        Should be same length as the list `f`, and need not be normalised.
+        Default: `numpy.ones_like(f)`
 
     weights: 1D array-like, optional
         sample weights (or list of weights), if desired. Should have length
-        same as samples.shape[0]
-        default: `numpy.ones_like(samples)`
+        same as `samples.shape[0]`.
+        Default: `numpy.ones_like(samples)`
 
     ny: int, optional
-        Resolution of y axis
-        default: `100`
+        Resolution of `y` axis.
+        Default: `100`
 
     y: array-like, optional
-        Explicit descriptor of y values to evaluate.
-        default: `numpy.linspace(min(f), max(f), ny)`
+        Explicit descriptor of `y` values to evaluate.
+        Default: `numpy.linspace(min(f), max(f), ny)`
 
     ntrim: int, optional
         Approximate number of samples to trim down to, if desired. Useful if
-        the posterior is dramatically oversampled
[…]
