### 2018-08-22 #849 @williamjameshandley
@zhampel Many thanks for your detailed comments.
* Paper: I think I've addressed all your issues.
* Documentation: I've tried to make the docstrings as descriptive as possible, without being too verbose. I agree the evidence was unclear. Hopefully now it's relatively obvious what these functions are expecting as an optional argument.
* Installation. Updated the readme accordingly. Does installing [Pillow](https://pypi.org/project/Pillow/) fix your issues? If so I'll update the Readme for mac-users.

Once you're happy with my updates (all on github), and we've fixed the OSX failure, I'll update the version numbers and PyPi.

> I really like the kwargs.pop functionality you put in. So smooth :)

I don't know how to do it any other way now. It came from [Effective Python](https://effectivepython.com/), which I thoroughly recommend (and is split up into nice, bitesize examples).
