### 2018-08-23 #849 @zhampel
@williamjameshandley

- For the paper edit, you are missing periods in the last sentence of the last paragraph, as well as one at the end of the caption of Figure 1.
- From your reply to the the documentation comments, it seemed like you were going to make edits to the docstrings. I didn't see any commits to that end. Were you planning on doing so or just agreeing that there was some ambiguity there?
- I will try the Pillow install today to check your fix.
