### 2018-08-24 #849 @zhampel
@williamjameshandley
- I must not have refreshed my page properly or something, so I finally saw the docstring edits you made.
- After installing pillow now, I get some ignored QueueManagerThread warning, but no error.

Good job cleaning this up!
