### 2018-08-24 #849 @zhampel
@williamjameshandley

Code
====
A few comments about the code. They are minor and nit-picky. Everything else looks well-written and functionally appropriate.

- _check_args in _utils.py: There are a few extra indents on some `raise ValueError` statements
- _normalize_weights in _utils.py: Could use a little more space for readability. It's a little tight compared to the rest of your code.

Maybe rerun flake8 just to check one last time to make sure there aren't any other stragglers.

One more thing about the testing suite:
- is test_plot_lines actually running a test? I don’t see an image_comparison style test like you have in test_plotting in test_drivers.py.
