### 2018-08-24 #849 @zhampel
@arfon I've checked off all my boxes and consider my review finished. I have made a few last minor comments to @williamjameshandley, though they are stylistic and do not require my approval for publication.
