### 2018-08-24 #849 @zhampel
@arfon Also, I can't find an email link on your personal websites. Can you please send me a quick email from the link provided on my own homepage [here](https://zhampel.github.io/)? Thanks.
