### 2018-08-24 #849 @zhampel
@arfon sorry that github email isn't working for some reason. My academic email zhampel@icecube.wisc.edu is fine please.
