### 2018-08-27 223a1bd
Paper updates
