### 2018-08-27 4bb7ffb
Pillow note to README.rst
