+++ fgivenx/_utils.py
@@ -28,8 +28,8 @@ def _check_args(logZ, f, x, samples, weights):
 
     # f
     if len(logZ) != len(f):
-            raise ValueError("len(logZ) = %i != len(f)= %i"
-                             % (len(logZ), len(f)))
+        raise ValueError("len(logZ) = %i != len(f)= %i"
+                         % (len(logZ), len(f)))
     for func in f:
         if not callable(func):
             raise ValueError("first argument f must be function"
@@ -37,8 +37,8 @@ def _check_args(logZ, f, x, samples, weights):
 
     # samples
     if len(logZ) != len(samples):
-            raise ValueError("len(logZ) = %i != len(samples)= %i"
-                             % (len(logZ), len(samples)))
+        raise ValueError("len(logZ) = %i != len(samples)= %i"
+                         % (len(logZ), len(samples)))
     samples = [numpy.array(s, dtype='double') for s in samples]
     for s in samples:
         if len(s.shape) is not 2:
@@ -46,8 +46,8 @@ def _check_args(logZ, f, x, samples, weights):
 
     # weights
     if len(logZ) != len(weights):
-            raise ValueError("len(logZ) = %i != len(weights)= %i"
-                             % (len(logZ), len(weights)))
+        raise ValueError("len(logZ) = %i != len(weights)= %i"
+                         % (len(logZ), len(weights)))
     weights = [numpy.array(w, dtype='double') if w is not None
                else numpy.ones(len(s), dtype='double')
                for w, s in zip(weights, samples)]
@@ -85,13 +85,20 @@ def _normalise_weights(logZ, weights, ntrim=None):
         normalised weights
     """
     logZ -= logZ.max()
+
     Zs = numpy.exp(logZ)
+
     weights = [w/w.sum()*Z for w, Z in zip(weights, Zs)]
+
     wmax = max([w.max() for w in weights])
+
     weights = [w/wmax for w in weights]
+
     ntot = sum([w.sum() for w in weights])
+
     if ntrim is not None and ntrim < ntot:
         weights = [w*ntrim/ntot for w in weights]
+
     return logZ, weights
 
 
