### 2018-08-28 c381803
Added better image comparison test
