### 2018-08-28 5e9e27d
Further test fixes
