### 2018-08-28 3249678
Paper updates
