+++ paper/paper.bib
@@ -110,12 +110,12 @@
 
 @article{zenodo,
     Author = {Handley, W},
-    doi = {10.5281/zenodo.1345486},
+    doi = {10.5281/zenodo.1404566},
     Month = {Aug},
     Publisher = {Zenodo},
-    Title = {fgivenx: v2.1.11},
+    Title = {fgivenx: v2.1.16},
     Year = {2018},
-    url = {http://dx.doi.org/10.5281/zenodo.1345486}
+    url = {http://dx.doi.org/10.5281/zenodo.1404566}
 }
 
 @unpublished{Higson2018b,
