+++ fgivenx/test/test_plot.py
@@ -89,4 +89,4 @@ def test_plot_lines_no_ax():
 def test_plot_lines_downsample():
     x, fsamps = gen_line_data()
     plt.subplots()
-    plot_lines(x, fsamps, downsample=50)
+    plot_lines(x, fsamps, downsample=200)
