### 2018-08-28 aaf4639
Paper version update
