+++ paper/paper.bib
@@ -110,12 +110,12 @@
 
 @article{zenodo,
     Author = {Handley, W},
-    doi = {10.5281/zenodo.1404566},
+    doi = {10.5281/zenodo.1404584},
     Month = {Aug},
     Publisher = {Zenodo},
-    Title = {fgivenx: v2.1.16},
+    Title = {fgivenx: v2.1.17},
     Year = {2018},
-    url = {http://dx.doi.org/10.5281/zenodo.1404566}
+    url = {http://dx.doi.org/10.5281/zenodo.1404584}
 }
 
 @unpublished{Higson2018b,
