### 2018-08-28 #849 @williamjameshandley
@zhampel

>* I must not have refreshed my page properly or something, so I finally saw the docstring edits you made.

I find one often needs to force refresh the cache in order to see updates

>Maybe rerun flake8 just to check one last time to make sure there aren't any other stragglers.

Done

>is test_plot_lines actually running a test? I don’t see an image_comparison style test like you have in test_plotting in test_drivers.py.

I've added some more tests to this, so that each separate test actually runs `image_comparison` tests. Tests and figures also uploaded to PyPi.

>After installing pillow now, I get some ignored QueueManagerThread warning, but no error.

Added a note to the README.rst

@arfon I believe that I have now addressed all the issues raised
